Separate DrawNodeSystem and NavigationSystem from drawgalaxy
Move these two classes to their own files.
Give them header files.
