+++ engine/CMakeLists.txt
@@ -397,6 +397,8 @@ SET(LIBGFXCLIENT_SOURCES
     src/gfx/nav/navpath.cpp
     src/gfx/nav/navscreen.cpp
     src/gfx/nav/navscreenoccupied.cpp
+    src/gfx/nav/system_draw_node.cpp
+    src/gfx/nav/navigation_system.cpp
     src/gfx/radar/bubble_display.cpp
     src/gfx/radar/bubble_display.h
     src/gfx/radar/dual_display.cpp
